requirements.txt
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-embeddings-bedrock
llama-index-llms-bedrock
sentence-transformers
llama-index
chromadb
pydantic
streamlit